% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericquickpredict.R
\name{genericquickpredict}
\alias{genericquickpredict}
\title{genericquickpredict takes a quickPredictResponse and returns prediction}
\usage{
genericquickpredict(independentFeatures, rawModel, implmentedWith)
}
\description{
genericquickpredict takes a quickPredictResponse and returns prediction
}
