% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictgenericlmglm.R
\name{jaqpot.predict.base.lm.glm}
\alias{jaqpot.predict.base.lm.glm}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
jaqpot.predict.base.lm.glm(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
