% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bnlearn_discrete.R
\name{jaqpot.predict.bnlearn.discrete}
\alias{jaqpot.predict.bnlearn.discrete}
\title{jaqpot.predict.bnlearn.discrete makes a PredictionResponse for Jaqpot for a discrete bayesian model created using package 'bnlearn'}
\usage{
jaqpot.predict.bnlearn.discrete(dataset, rawModel, additionalInfo)
}
\arguments{
\item{dataset}{The input from the user along with the feature names.}

\item{rawModel}{The serialised model.}

\item{additionalInfo}{additional information from user e.g. prediction method specification.}
}
\description{
jaqpot.predict.bnlearn.discrete makes a PredictionResponse for Jaqpot for a discrete bayesian model created using package 'bnlearn'
}
