% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_caret.R
\name{jaqpot.predict.caret}
\alias{jaqpot.predict.caret}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
jaqpot.predict.caret(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
