% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gbm.R
\name{jaqpot.predict.gbm}
\alias{jaqpot.predict.gbm}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
jaqpot.predict.gbm(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
