% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_naive_bayess.R
\name{jaqpot.predict.naivebayess}
\alias{jaqpot.predict.naivebayess}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
jaqpot.predict.naivebayess(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
