% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpart.R
\name{jaqpot.predict.rpart}
\alias{jaqpot.predict.rpart}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
jaqpot.predict.rpart(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
