% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tree_regr.R
\name{jaqpot.predict.tree}
\alias{jaqpot.predict.tree}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
jaqpot.predict.tree(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
