% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictgenericlmglm.R
\name{predict.base.lm.glm}
\alias{predict.base.lm.glm}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
\method{predict}{base.lm.glm}(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
