% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpart.R
\name{predict.rpart.jaqpot}
\alias{predict.rpart.jaqpot}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
\method{predict}{rpart.jaqpot}(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
