% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tree_regr.R
\name{predict.tree.regr}
\alias{predict.tree.regr}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
\method{predict}{tree.regr}(dataset, rawModel, additionalInfo)
}
\arguments{
\item{additionalInfo}{}
}
\description{
predict makes a PredictionResponse for Jaqpot
}
