% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictgeneric.R
\name{predictgeneric.lm.glm}
\alias{predictgeneric.lm.glm}
\title{predict makes a PredictionResponse for Jaqpot}
\usage{
predictgeneric.lm.glm(dataset, rawModel, additionalInfo)
}
\description{
predict makes a PredictionResponse for Jaqpot
}
